ROC:2485 - Refactoring file name
